+++ src/utils.jl
@@ -8,3 +8,4 @@ function _hex_escape(s::String)
     strs
 end
 _hex_escape(s) = _hex_escape(string(s))
+